+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Use paginated results for the newsfeed (#524)
 -   Newsfeed card icon for canceled compute plans (#525)
 
+### Fixed
+
+-   synced state init issue (#543)
+
 ## [0.19.0] - 2022-04-19
 
 ## Changed